+++ Source/Core/Common.h
@@ -17,8 +17,8 @@
 std::string const RelativePath(std::string const &aFileName);
 
 #ifndef _WIN32
-float GetTickCount();
-float timeGetTime();
+long GetTickCount();
+long timeGetTime();
 #endif
 
 int StringToInt(std::string const &value);
